pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/shift/version.py"
pattern = "VERSION = \"(?P<version>[^\"]+)\""

[project]
name = "nrel-shift"
dynamic = ["version"]
description = "Framework for developing synthetic distribution feeder model."
readme = "README.md"
requires-python = ">=3.10"
license = "BSD-3-clause"
keywords = []
authors = [
  { name = "Kapil Duwadi", email = "Kapil.Duwadi@nrel.gov" },
  { name = "Aadil Latif", email = "Aadil.Latif@nrel.gov" },
  { name = "Erik Pohl", email = "Erik.Pohl@nrel.gov" },
]
classifiers = [
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
]
dependencies = [
  "osmnx",
  "scikit-learn",
  "plotly",
  "geopy",
  "grid-data-models==2.2.1",
  "importlib-metadata",
  "loguru",
]

[project.optional-dependencies]
dev = ["pre-commit", "pytest", "pytest-cov", "pytest-mock", "ruff==0.14.14"]
doc = ["sphinx", "pydata-sphinx-theme", "myst-parser", "autodoc_pydantic"]
mcp = ["mcp[cli]>=1.2.0", "pyyaml"]

[project.scripts]
shift-mcp-server = "shift.mcp_server.__main__:main"

[project.urls]
Documentation = "https://github.com/NREL-Distribution-Suites/shift#readme"
Issues = "https://github.com/NREL-Distribution-Suites/shift/issues"
Source = "https://github.com/NREL-Distribution-Suites/shift"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".git",
  ".ruff_cache",
  ".venv",
  "_build",
  "build",
  "dist",
  "env",
  "venv",
]
line-length = 99
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
  "C901", # McCabe complexity
  "E4",   # Subset of pycodestyle (E)
  "E7",
  "E9",
  "F",    # Pyflakes
  "W",    # pycodestyle warnings
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.hatch.build.targets.wheel]
packages = ["src/shift"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/shift"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true
